move everything under examples
