+++ examples/agents/inflation.py
