+++ examples/agents/rag_as_attachments.py
