+++ examples/memory/client.py
